Don't execute redirect for a child action (#17)
When a view contains a child action: 

Html.RenderAction("actionName", "controllerName")

and the filter causes a redirect, an MVC error is thrown, "child actions are not allowed to perform redirect actions". This is because the filter looks at the attributes of the currently executing child controller, not the parent controller. If both don't have [NoLowercaseQueryString], the exception may be thrown.

Fix is to ignore execution of the redirect filter during child action execution.